refactor: extract menus into js/menus.js with callback injection
