+++ js/app.js
@@ -108,6 +108,12 @@ import {
   generateThumbnails, highlightActiveThumbnail,
   renderThumbnailForItem, setThumbnailCallbacks,
 } from './thumbnails.js';
+import {
+  closeDropdown, openDropdown, getMenuDefinitions,
+  initDropdownMenus, buildRecentSubmenu,
+  showContextMenu, hideContextMenu, showAnnotationContextMenu,
+  setMenuCallbacks,
+} from './menus.js';
 
 /* ═══════════════════ Initialization ═══════════════════ */
 
@@ -147,7 +153,20 @@ async function boot() {
 
     // Wire up all UI events
     wireEvents();
-    initDropdownMenus();
+
+    // Register menu callbacks for context menus (before initDropdownMenus)
+    setMenuCallbacks({
+      // Context menu callbacks
+      showLoading, hideLoading, toast, reloadAfterEdit,
+      insertBlankPage, handleAddPages, rotatePage, deletePage,
+      splitPDF, downloadBlob,
+      // Annotation context menu callbacks
+      getCanvas, copySelected, pasteClipboard, duplicateSelected,
+      bringToFront, sendToBack, bringForward, sendBackward,
+      lockSelected, unlockSelected, deleteSelected,
+    });
+
+    initDropdownMenus(_getMenuActions());
     initMenuActions();
 
     // Listen for actual size event from menu-actions
@@ -1157,249 +1176,7 @@ async function loadBookmarks() {
   }
 }
 
-/* ═══════════════════ Thumbnail Context Menu ═══════════════════ */
-
-let contextMenu = null;
-
-function showContextMenu(e, pageNum) {
-  e.preventDefault();
-  hideContextMenu();
-
-  contextMenu = document.createElement('div');
-  contextMenu.className = 'context-menu';
-  contextMenu.innerHTML = `
-    <button data-action="insert-before">${icon('file-plus', 14)} Insert Page Before</button>
-    <button data-action="insert-after">${icon('file-plus', 14)} Insert Page After</button>
-    <button data-action="insert-blank">${icon('file-plus', 14)} Insert Blank Page</button>
-    <button data-action="duplicate">${icon('files', 14)} Duplicate Page</button>
-    <div class="context-menu-separator"></div>
-    <button data-action="rotate-cw">${icon('rotate-cw', 14)} Rotate Right</button>
-    <button data-action="rotate-ccw">${icon('rotate-ccw', 14)} Rotate Left</button>
-    <button data-action="rotate-180">${icon('flip-vertical', 14)} Rotate 180°</button>
-    <div class="context-menu-separator"></div>
-    <button data-action="delete" ${State.totalPages <= 1 ? 'disabled' : ''}>${icon('trash', 14)} Delete Page</button>
-    <button data-action="extract">${icon('file-output', 14)} Extract Page</button>
-  `;
-
-  // Position at mouse
-  contextMenu.style.left = e.clientX + 'px';
-  contextMenu.style.top = e.clientY + 'px';
-  document.body.appendChild(contextMenu);
-
-  // Clamp to viewport
-  const rect = contextMenu.getBoundingClientRect();
-  if (rect.right > window.innerWidth) {
-    contextMenu.style.left = (window.innerWidth - rect.width - 8) + 'px';
-  }
-  if (rect.bottom > window.innerHeight) {
-    contextMenu.style.top = (window.innerHeight - rect.height - 8) + 'px';
-  }
-
-  // Handle clicks
-  contextMenu.addEventListener('click', async (ev) => {
-    const btn = ev.target.closest('[data-action]');
-    if (!btn || btn.disabled) return;
-    hideContextMenu();
-
-    const action = btn.dataset.action;
-    const idx = pageNum - 1; // 0-based
-
-    // Insert blank page
-    if (action === 'insert-blank') {
-      showLoading('Inserting page…');
-      try {
-        const newBytes = await insertBlankPage(State.pdfBytes, idx);
-        State.currentPage = pageNum + 1; // navigate to the new blank page
-        await reloadAfterEdit(newBytes);
-        toast('Inserted blank page', 'success');
-      } catch (err) {
-        console.error('Insert blank page failed:', err);
-        toast('Insert blank page failed: ' + err.message, 'error');
-      } finally {
-        hideLoading();
-      }
-      return;
-    }
-
-    // Insert page needs a file picker — not a loading operation
-    if (action === 'insert-before' || action === 'insert-after') {
-      const input = document.createElement('input');
-      input.type = 'file';
-      input.accept = '.pdf';
-      input.multiple = true;
-      input.addEventListener('change', e => {
-        const files = Array.from(e.target.files);
-        if (!files.length) return;
-        const insertAfter = action === 'insert-before' ? idx - 1 : idx;
-        handleAddPages(files, insertAfter);
-      });
-      input.click();
-      return;
-    }
-
-    // Duplicate page via pdf-lib
-    if (action === 'duplicate') {
-      showLoading('Duplicating page…');
-      try {
-        const PDFLib = window.PDFLib;
-        const doc = await PDFLib.PDFDocument.load(State.pdfBytes, { ignoreEncryption: true });
-        const [copied] = await doc.copyPages(doc, [idx]);
-        doc.insertPage(idx + 1, copied);
-        const newBytes = await doc.save();
-        State.currentPage = pageNum + 1;
-        await reloadAfterEdit(newBytes);
-        toast(`Duplicated page ${pageNum}`, 'success');
-      } catch (err) {
-        console.error('Duplicate page failed:', err);
-        toast('Duplicate failed: ' + err.message, 'error');
-      } finally {
-        hideLoading();
-      }
-      return;
-    }
-
-    showLoading('Editing page…');
-    try {
-      let newBytes;
-      switch (action) {
-        case 'rotate-cw':
-          newBytes = await rotatePage(State.pdfBytes, idx, 90);
-          await reloadAfterEdit(newBytes);
-          toast('Rotated page right', 'success');
-          break;
-        case 'rotate-ccw':
-          newBytes = await rotatePage(State.pdfBytes, idx, -90);
-          await reloadAfterEdit(newBytes);
-          toast('Rotated page left', 'success');
-          break;
-        case 'rotate-180':
-          newBytes = await rotatePage(State.pdfBytes, idx, 180);
-          await reloadAfterEdit(newBytes);
-          toast('Rotated page 180°', 'success');
-          break;
-        case 'delete':
-          if (State.totalPages <= 1) return;
-          if (!confirm(`Delete page ${pageNum}? This cannot be undone.`)) return;
-          newBytes = await deletePage(State.pdfBytes, idx);
-          // If we deleted the current page or a page before it, adjust
-          if (pageNum <= State.currentPage && State.currentPage > 1) {
-            State.currentPage--;
-          }
-          await reloadAfterEdit(newBytes);
-          toast(`Deleted page ${pageNum}`, 'success');
-          break;
-        case 'extract': {
-          const extracted = await splitPDF(State.pdfBytes, [[idx]]);
-          const part = extracted[0];
-          const name = State.fileName.replace('.pdf', '') + `_${part.label}.pdf`;
-          downloadBlob(part.bytes, name);
-          toast(`Extracted ${part.label}`, 'success');
-          break;
-        }
-      }
-    } catch (err) {
-      console.error('Page operation failed:', err);
-      toast('Operation failed: ' + err.message, 'error');
-    } finally {
-      hideLoading();
-    }
-  });
-
-  // Close on click outside or Escape
-  setTimeout(() => {
-    document.addEventListener('click', hideContextMenu, { once: true });
-    document.addEventListener('keydown', function onKey(ev) {
-      if (ev.key === 'Escape') {
-        hideContextMenu();
-        document.removeEventListener('keydown', onKey);
-      }
-    });
-  }, 0);
-}
-
-function hideContextMenu() {
-  if (contextMenu) {
-    contextMenu.remove();
-    contextMenu = null;
-  }
-}
-
-/* ═══════════════════ Annotation Context Menu ═══════════════════ */
-
-function showAnnotationContextMenu(e, target) {
-  e.preventDefault();
-  e.stopPropagation();
-  hideContextMenu();
-
-  const canvas = getCanvas();
-  if (!canvas) return;
-
-  const locked = target ? !!target.lockMovementX : false;
-  const hasSelection = !!target;
-
-  contextMenu = document.createElement('div');
-  contextMenu.className = 'context-menu';
-  contextMenu.innerHTML = `
-    <button data-action="anno-copy" ${!hasSelection ? 'disabled' : ''}>${icon('copy', 14)} Copy</button>
-    <button data-action="anno-paste">${icon('clipboard-paste', 14)} Paste</button>
-    <button data-action="anno-duplicate" ${!hasSelection ? 'disabled' : ''}>${icon('files', 14)} Duplicate</button>
-    <div class="context-menu-separator"></div>
-    <button data-action="anno-front" ${!hasSelection ? 'disabled' : ''}>${icon('arrow-up-to-line', 14)} Bring to Front</button>
-    <button data-action="anno-forward" ${!hasSelection ? 'disabled' : ''}>${icon('chevron-up', 14)} Bring Forward</button>
-    <button data-action="anno-backward" ${!hasSelection ? 'disabled' : ''}>${icon('chevron-down', 14)} Send Backward</button>
-    <button data-action="anno-back" ${!hasSelection ? 'disabled' : ''}>${icon('arrow-down-to-line', 14)} Send to Back</button>
-    <div class="context-menu-separator"></div>
-    <button data-action="anno-lock" ${!hasSelection ? 'disabled' : ''}>${locked ? icon('lock', 14) + ' Unlock' : icon('lock-open', 14) + ' Lock'}</button>
-    <button data-action="anno-delete" ${!hasSelection ? 'disabled' : ''}>${icon('trash', 14)} Delete</button>
-  `;
-
-  // Position at mouse
-  contextMenu.style.left = e.clientX + 'px';
-  contextMenu.style.top = e.clientY + 'px';
-  document.body.appendChild(contextMenu);
-
-  // Clamp to viewport
-  const rect = contextMenu.getBoundingClientRect();
-  if (rect.right > window.innerWidth) {
-    contextMenu.style.left = (window.innerWidth - rect.width - 8) + 'px';
-  }
-  if (rect.bottom > window.innerHeight) {
-    contextMenu.style.top = (window.innerHeight - rect.height - 8) + 'px';
-  }
-
-  // Handle clicks
-  contextMenu.addEventListener('click', (ev) => {
-    const btn = ev.target.closest('[data-action]');
-    if (!btn || btn.disabled) return;
-    hideContextMenu();
-
-    switch (btn.dataset.action) {
-      case 'anno-copy':      copySelected(); break;
-      case 'anno-paste':     pasteClipboard(); break;
-      case 'anno-duplicate': duplicateSelected(); break;
-      case 'anno-front':     bringToFront(); break;
-      case 'anno-forward':   bringForward(); break;
-      case 'anno-backward':  sendBackward(); break;
-      case 'anno-back':      sendToBack(); break;
-      case 'anno-lock':
-        if (locked) unlockSelected();
-        else lockSelected();
-        break;
-      case 'anno-delete':    deleteSelected(); break;
-    }
-  });
-
-  // Close on click outside or Escape
-  setTimeout(() => {
-    document.addEventListener('click', hideContextMenu, { once: true });
-    document.addEventListener('keydown', function onKey(ev) {
-      if (ev.key === 'Escape') {
-        hideContextMenu();
-        document.removeEventListener('keydown', onKey);
-      }
-    });
-  }, 0);
-}
+/* Context menu functions extracted to js/menus.js */
 
 /* ═══════════════════ Merge Modal ═══════════════════ */
 
@@ -3016,287 +2793,7 @@ function initModalFocusTrapping() {
   });
 }
 
-/* ═══════════════════ Title Bar Dropdown Menus ═══════════════════ */
-
-let _activeDropdown = null;
-
-function closeDropdown() {
-  if (_activeDropdown) {
-    if (_activeDropdown._keyHandler) {
-      document.removeEventListener('keydown', _activeDropdown._keyHandler);
-    }
-    _activeDropdown.remove();
-    _activeDropdown = null;
-  }
-  // Remove any lingering submenus
-  document.querySelectorAll('.dropdown-submenu').forEach(s => s.remove());
-  document.querySelectorAll('.menu-item.active').forEach(el => {
-    el.classList.remove('active');
-    el.setAttribute('aria-expanded', 'false');
-  });
-}
-
-function openDropdown(menuBtn, items) {
-  closeDropdown();
-  hideContextMenu();
-
-  menuBtn.classList.add('active');
-  menuBtn.setAttribute('aria-expanded', 'true');
-  const rect = menuBtn.getBoundingClientRect();
-  const menu = document.createElement('div');
-  menu.className = 'dropdown-menu';
-  menu.setAttribute('role', 'menu');
-
-  for (const item of items) {
-    if (item === '---') {
-      const sep = document.createElement('div');
-      sep.className = 'dropdown-menu-separator';
-      sep.setAttribute('role', 'separator');
-      menu.appendChild(sep);
-      continue;
-    }
-    const btn = document.createElement('button');
-    btn.setAttribute('role', 'menuitem');
-    const needsDoc = item.needsDoc !== false;
-    const isDisabled = (needsDoc && !State.pdfBytes) || item.disabled;
-    btn.disabled = isDisabled;
-
-    // Submenu arrow indicator
-    const submenuArrow = item.submenu ? `<span class="submenu-arrow">${icon('chevron-right', 12)}</span>` : '';
-    const shortcutHtml = item.shortcut && !item.submenu ? `<span class="shortcut-hint">${item.shortcut}</span>` : '';
-
-    btn.innerHTML = `${icon(item.icon, 14)}<span>${item.label}</span>${shortcutHtml}${submenuArrow}`;
-
-    if (item.submenu) {
-      // Submenu on hover
-      let submenuEl = null;
-      let submenuTimeout = null;
-      btn.addEventListener('mouseenter', () => {
-        clearTimeout(submenuTimeout);
-        // Remove any other open submenus
-        menu.querySelectorAll('.dropdown-submenu').forEach(s => s.remove());
-        const subItems = item.submenu();
-        submenuEl = document.createElement('div');
-        submenuEl.className = 'dropdown-menu dropdown-submenu';
-        submenuEl.setAttribute('role', 'menu');
-        for (const si of subItems) {
-          if (si === '---') {
-            const sep = document.createElement('div');
-            sep.className = 'dropdown-menu-separator';
-            submenuEl.appendChild(sep);
-            continue;
-          }
-          const subBtn = document.createElement('button');
-          subBtn.setAttribute('role', 'menuitem');
-          const subNeedsDoc = si.needsDoc !== false;
-          subBtn.disabled = (subNeedsDoc && !State.pdfBytes) || si.disabled;
-          subBtn.innerHTML = `${icon(si.icon, 14)}<span>${si.label}</span>${si.shortcut ? `<span class="shortcut-hint">${si.shortcut}</span>` : ''}`;
-          subBtn.addEventListener('click', () => {
-            closeDropdown();
-            si.action();
-          });
-          submenuEl.appendChild(subBtn);
-        }
-        const btnRect = btn.getBoundingClientRect();
-        submenuEl.style.left = btnRect.right + 'px';
-        submenuEl.style.top = btnRect.top + 'px';
-        document.body.appendChild(submenuEl);
-        // Keep submenu in viewport
-        const subRect = submenuEl.getBoundingClientRect();
-        if (subRect.right > window.innerWidth) {
-          submenuEl.style.left = (btnRect.left - subRect.width) + 'px';
-        }
-        if (subRect.bottom > window.innerHeight) {
-          submenuEl.style.top = (window.innerHeight - subRect.height - 8) + 'px';
-        }
-      });
-      btn.addEventListener('mouseleave', (e) => {
-        submenuTimeout = setTimeout(() => {
-          if (submenuEl && !submenuEl.matches(':hover')) {
-            submenuEl.remove();
-            submenuEl = null;
-          }
-        }, 200);
-      });
-    } else {
-      btn.addEventListener('click', () => {
-        closeDropdown();
-        item.action();
-      });
-    }
-    menu.appendChild(btn);
-  }
-
-  menu.style.left = rect.left + 'px';
-  menu.style.top = rect.bottom + 'px';
-  document.body.appendChild(menu);
-
-  // Keep menu within viewport
-  const menuRect = menu.getBoundingClientRect();
-  if (menuRect.right > window.innerWidth) {
-    menu.style.left = (window.innerWidth - menuRect.width - 8) + 'px';
-  }
-
-  _activeDropdown = menu;
-
-  // Focus first enabled button for keyboard navigation
-  const firstBtn = menu.querySelector('button:not(:disabled)');
-  if (firstBtn) firstBtn.focus();
-
-  // Close on click outside
-  const onOutsideClick = (e) => {
-    if (!menu.contains(e.target) && !e.target.closest('.menu-item') && !e.target.closest('.dropdown-submenu')) {
-      closeDropdown();
-      document.removeEventListener('mousedown', onOutsideClick);
-    }
-  };
-  setTimeout(() => document.addEventListener('mousedown', onOutsideClick), 0);
-
-  // Keyboard navigation: Arrow keys, Enter, Escape
-  const onKey = (e) => {
-    const buttons = Array.from(menu.querySelectorAll('button:not(:disabled)'));
-    const focused = document.activeElement;
-    const idx = buttons.indexOf(focused);
-
-    if (e.key === 'Escape') {
-      closeDropdown();
-      menuBtn.focus();
-      document.removeEventListener('keydown', onKey);
-      e.preventDefault();
-    } else if (e.key === 'ArrowDown') {
-      e.preventDefault();
-      const next = idx < buttons.length - 1 ? idx + 1 : 0;
-      buttons[next]?.focus();
-    } else if (e.key === 'ArrowUp') {
-      e.preventDefault();
-      const prev = idx > 0 ? idx - 1 : buttons.length - 1;
-      buttons[prev]?.focus();
-    } else if (e.key === 'ArrowRight') {
-      // Open submenu if item has one, or move to next menu
-      const item = items.filter(i => i !== '---')[idx];
-      if (item && item.submenu) {
-        // Trigger mouseenter to open submenu
-        focused.dispatchEvent(new MouseEvent('mouseenter'));
-        setTimeout(() => {
-          const sub = document.querySelector('.dropdown-submenu button:not(:disabled)');
-          if (sub) sub.focus();
-        }, 50);
-        e.preventDefault();
-      } else {
-        // Move to next menu item in title bar
-        const allMenuBtns = Array.from(document.querySelectorAll('.menu-item'));
-        const menuIdx = allMenuBtns.indexOf(menuBtn);
-        if (menuIdx < allMenuBtns.length - 1) {
-          allMenuBtns[menuIdx + 1].click();
-          e.preventDefault();
-        }
-      }
-    } else if (e.key === 'ArrowLeft') {
-      // Move to previous menu item in title bar
-      const allMenuBtns = Array.from(document.querySelectorAll('.menu-item'));
-      const menuIdx = allMenuBtns.indexOf(menuBtn);
-      if (menuIdx > 0) {
-        allMenuBtns[menuIdx - 1].click();
-        e.preventDefault();
-      }
-    } else if (e.key === 'Home') {
-      e.preventDefault();
-      buttons[0]?.focus();
-    } else if (e.key === 'End') {
-      e.preventDefault();
-      buttons[buttons.length - 1]?.focus();
-    }
-  };
-  document.addEventListener('keydown', onKey);
-  // Store cleanup reference
-  menu._keyHandler = onKey;
-}
-
-function getMenuDefinitions() {
-  return {
-    'File': [
-      { icon: 'file-plus', label: 'New Blank PDF', needsDoc: false, action: handleNewBlankPdf },
-      { icon: 'folder-open', label: 'Open', shortcut: 'Ctrl+O', needsDoc: false, action: () => DOM.fileInput.click() },
-      { icon: 'clock', label: 'Open Recent', needsDoc: false, submenu: () => buildRecentSubmenu() },
-      '---',
-      { icon: 'save', label: 'Save', shortcut: 'Ctrl+S', action: handleSave },
-      { icon: 'download', label: 'Save & Download', shortcut: 'Ctrl+Shift+S', action: handleSaveDownload },
-      { icon: 'file-output', label: 'Export…', action: handleExport },
-      { icon: 'download', label: 'Export as Image', action: () => $('btn-export-image').click() },
-      '---',
-      { icon: 'printer', label: 'Print', shortcut: 'Ctrl+P', action: handlePrint },
-      { icon: 'info', label: 'Properties', action: () => togglePropertiesPanel(true) },
-      { icon: 'x', label: 'Close', action: handleCloseDocument },
-    ],
-    'Edit': [
-      { icon: 'undo', label: 'Undo', shortcut: 'Ctrl+Z', action: handleUndo },
-      { icon: 'redo', label: 'Redo', shortcut: 'Ctrl+Shift+Z', action: handleRedo },
-      '---',
-      { icon: 'scissors', label: 'Cut', shortcut: 'Ctrl+X', action: () => document.execCommand('cut') },
-      { icon: 'copy', label: 'Copy', shortcut: 'Ctrl+C', action: () => copySelected() },
-      { icon: 'clipboard-paste', label: 'Paste', shortcut: 'Ctrl+V', action: () => pasteClipboard() },
-      { icon: 'trash', label: 'Delete', shortcut: 'Del', action: deleteSelected },
-      '---',
-      { icon: 'maximize', label: 'Select All', shortcut: 'Ctrl+A', action: () => { const c = getCanvas(); if (c) { c.discardActiveObject(); const sel = new fabric.ActiveSelection(c.getObjects(), { canvas: c }); c.setActiveObject(sel); c.requestRenderAll(); } } },
-      '---',
-      { icon: 'search', label: 'Find & Replace', shortcut: 'Ctrl+F', action: () => openFindBar(true) },
-    ],
-    'View': [
-      { icon: 'zoom-in', label: 'Zoom In', shortcut: 'Ctrl+=', action: zoomIn },
-      { icon: 'zoom-out', label: 'Zoom Out', shortcut: 'Ctrl+\u2212', action: zoomOut },
-      '---',
-      { icon: 'columns', label: 'Fit Width', action: fitWidth },
-      { icon: 'maximize', label: 'Fit Page', action: fitPage },
-      { icon: 'scan', label: 'Actual Size', shortcut: 'Ctrl+0', action: () => setZoom(1.0) },
-      '---',
-      { icon: 'panel-left-open', label: 'Toggle Sidebar', needsDoc: false, action: toggleSidebar },
-      { icon: 'panel-right-open', label: 'Toggle Properties Panel', action: () => togglePropertiesPanel() },
-      '---',
-      { icon: 'moon', label: 'Dark Mode', needsDoc: false, action: toggleDarkMode },
-      { icon: 'maximize', label: 'Full Screen', shortcut: 'F11', needsDoc: false, action: () => { if (!document.fullscreenElement) document.documentElement.requestFullscreen().catch(() => {}); else document.exitFullscreen(); } },
-    ],
-    'Insert': [
-      { icon: 'file-plus', label: 'Blank Page', action: () => $('btn-add-page').click() },
-      { icon: 'files', label: 'Pages from File', action: openMergeModal },
-      '---',
-      { icon: 'image', label: 'Image', action: handleImageInsert },
-      { icon: 'type', label: 'Text Annotation', shortcut: 'T', action: () => setTool('text') },
-      { icon: 'stamp', label: 'Stamp', action: () => setTool('stamp') },
-      '---',
-      { icon: 'pen-tool', label: 'Signature', action: () => $('btn-signature').click() },
-      { icon: 'droplet', label: 'Watermark', action: openWatermarkModal },
-      { icon: 'align-justify', label: 'Header / Footer', action: openHfModal },
-      '---',
-      { icon: 'hash', label: 'Bates Numbers', action: openBatesModal },
-      { icon: 'tag', label: 'Page Labels', action: openPageLabelsModal },
-      { icon: 'gavel', label: 'Exhibit Stamps', action: openExhibitModal },
-    ],
-    'Tools': [
-      { icon: 'type', label: 'Text Edit Mode', action: handleEditText },
-      '---',
-      { icon: 'form-input', label: 'Form Filler', action: () => { const btn = document.querySelector('[data-panel="forms"]'); if (btn) btn.click(); $('btn-detect-fields').click(); } },
-      { icon: 'list-ordered', label: 'Form Creator', action: () => { const btn = document.querySelector('[data-panel="forms"]'); if (btn) btn.click(); } },
-      '---',
-      { icon: 'file-scan', label: 'OCR', action: () => $('btn-ocr').click() },
-      { icon: 'shield-off', label: 'Redaction Patterns', action: () => $('btn-redact-search').click() },
-      { icon: 'git-compare', label: 'Document Compare', action: () => $('btn-compare').click() },
-      '---',
-      { icon: 'lock', label: 'Security', action: () => $('btn-encrypt').click() },
-      { icon: 'message-square', label: 'Comment Summary', action: openCommentSummaryModal },
-      '---',
-      { icon: 'maximize', label: 'Normalize Page Sizes', action: openNormalizePagesModal },
-      '---',
-      { icon: 'zap', label: 'Optimize / Compress', action: () => { $('optimize-result').textContent = ''; $('optimize-result').classList.add('hidden'); $('optimize-modal-backdrop').classList.remove('hidden'); } },
-    ],
-    'Help': [
-      { icon: 'info', label: 'Keyboard Shortcuts', shortcut: '?', needsDoc: false, action: openShortcutsModal },
-      { icon: 'play', label: 'Start Tour', needsDoc: false, action: () => window.__mbStartTour && window.__mbStartTour() },
-      '---',
-      { icon: 'zap', label: 'About Mudbrick', needsDoc: false, action: openAboutModal },
-      { icon: 'link', label: 'GitHub Repository', needsDoc: false, action: () => window.open('https://github.com/niclas-nicol/mudbrick', '_blank') },
-    ],
-  };
-}
+/* Dropdown menu functions extracted to js/menus.js */
 
 function handleNewBlankPdf() {
   ensurePdfLib().then(async () => {
@@ -3335,22 +2832,6 @@ function handleCloseDocument() {
   toast('Document closed');
 }
 
-function buildRecentSubmenu() {
-  const recent = getRecentFiles();
-  if (recent.length === 0) {
-    return [{ icon: 'info', label: 'No recent files', needsDoc: false, disabled: true, action: () => {} }];
-  }
-  const items = recent.map(f => ({
-    icon: 'file',
-    label: f.name.length > 35 ? f.name.slice(0, 32) + '\u2026' : f.name,
-    needsDoc: false,
-    action: () => DOM.fileInput.click(),
-  }));
-  items.push('---');
-  items.push({ icon: 'trash', label: 'Clear Recent', needsDoc: false, action: () => { clearRecentFiles(); toast('Recent files cleared'); } });
-  return items;
-}
-
 function openShortcutsModal() {
   $('shortcuts-modal-backdrop').classList.remove('hidden');
 }
@@ -3359,27 +2840,66 @@ function openAboutModal() {
   $('about-modal-backdrop').classList.remove('hidden');
 }
 
-function initDropdownMenus() {
-  const defs = getMenuDefinitions();
-  document.querySelectorAll('.menu-item').forEach(btn => {
-    const label = btn.textContent.trim();
-    const items = defs[label];
-    if (!items) return;
-    btn.addEventListener('click', (e) => {
-      e.stopPropagation();
-      if (_activeDropdown && btn.classList.contains('active')) {
-        closeDropdown();
-      } else {
-        openDropdown(btn, items);
-      }
-    });
-    // Open on hover when another menu is already open
-    btn.addEventListener('mouseenter', () => {
-      if (_activeDropdown) {
-        openDropdown(btn, items);
-      }
-    });
-  });
+/**
+ * Build the actions object for getMenuDefinitions / initDropdownMenus.
+ * Maps action names to the app-level functions they invoke.
+ */
+function _getMenuActions() {
+  return {
+    handleNewBlankPdf,
+    openFile: () => DOM.fileInput.click(),
+    getRecentFiles,
+    clearRecentFiles: () => { clearRecentFiles(); toast('Recent files cleared'); },
+    handleSave,
+    handleSaveDownload,
+    handleExport,
+    exportAsImage: () => $('btn-export-image').click(),
+    handlePrint,
+    showProperties: () => togglePropertiesPanel(true),
+    handleCloseDocument,
+    handleUndo,
+    handleRedo,
+    cut: () => document.execCommand('cut'),
+    copy: () => copySelected(),
+    paste: () => pasteClipboard(),
+    deleteSelected,
+    selectAll: () => { const c = getCanvas(); if (c) { c.discardActiveObject(); const sel = new fabric.ActiveSelection(c.getObjects(), { canvas: c }); c.setActiveObject(sel); c.requestRenderAll(); } },
+    openFindBar: () => openFindBar(true),
+    zoomIn,
+    zoomOut,
+    fitWidth,
+    fitPage,
+    actualSize: () => setZoom(1.0),
+    toggleSidebar,
+    togglePropertiesPanel: () => togglePropertiesPanel(),
+    toggleDarkMode,
+    fullScreen: () => { if (!document.fullscreenElement) document.documentElement.requestFullscreen().catch(() => {}); else document.exitFullscreen(); },
+    addBlankPage: () => $('btn-add-page').click(),
+    openMergeModal,
+    handleImageInsert,
+    setToolText: () => setTool('text'),
+    setToolStamp: () => setTool('stamp'),
+    openSignature: () => $('btn-signature').click(),
+    openWatermarkModal,
+    openHfModal,
+    openBatesModal,
+    openPageLabelsModal,
+    openExhibitModal,
+    handleEditText,
+    formFiller: () => { const btn = document.querySelector('[data-panel="forms"]'); if (btn) btn.click(); $('btn-detect-fields').click(); },
+    formCreator: () => { const btn = document.querySelector('[data-panel="forms"]'); if (btn) btn.click(); },
+    ocr: () => $('btn-ocr').click(),
+    redactSearch: () => $('btn-redact-search').click(),
+    docCompare: () => $('btn-compare').click(),
+    security: () => $('btn-encrypt').click(),
+    openCommentSummaryModal,
+    openNormalizePagesModal,
+    optimize: () => { $('optimize-result').textContent = ''; $('optimize-result').classList.add('hidden'); $('optimize-modal-backdrop').classList.remove('hidden'); },
+    openShortcutsModal,
+    startTour: () => window.__mbStartTour && window.__mbStartTour(),
+    openAboutModal,
+    openGitHub: () => window.open('https://github.com/niclas-nicol/mudbrick', '_blank'),
+  };
 }
 
 /* ═══════════════════ Modal Focus Management Helpers ═══════════════════ */
@@ -5792,12 +5312,12 @@ function handleKeyboard(e) {
     // Escape: close dropdown → close context menu → close modal → close crop
     //         → close find bar → deselect → switch to select tool
     case e.key === 'Escape': {
-      if (_activeDropdown) {
+      if (document.querySelector('.dropdown-menu')) {
         // Close title-bar dropdown and restore focus to the menu button
         const activeMenuBtn = document.querySelector('.menu-item.active');
         closeDropdown();
         if (activeMenuBtn) activeMenuBtn.focus();
-      } else if (contextMenu) {
+      } else if (document.querySelector('.context-menu')) {
         // Close thumbnail / annotation context menu
         hideContextMenu();
       } else {
